[Feat] #255 - Merge AddLinkVC & SelectClipVC
