+++ TOASTER-iOS/Application/Coordinator/Coordinators/HomeCoordinator.swift
@@ -46,7 +46,7 @@ private extension HomeCoordinator {
             self?.showDetailClipVC(id: id, name: name)
         }
         vc.onAddLinkSelected = { [weak self] in
-            //self?.startAddLinkCoordinator()
+            self?.showAddLinkVC()
         }
         router.setRoot(vc, animated: false)
     }
@@ -74,6 +74,16 @@ private extension HomeCoordinator {
         vc.onChangeRoot = { [weak self] in
             self?.router.dismiss()  // 로그아웃 완료 Alert dismiss
             self?.onFinish?()
+        }
+        router.push(vc, animated: true, hideBottomBarWhenPushed: true)
+    }
+    
+    func showAddLinkVC() {
+        let vc = viewControllerFactory.makeAddLinkVC(isNavigationBarHidden: true)
+        vc.onPopToRoot = { [weak self] in
+            self?.router.dismiss(animated: false) {
+                self?.router.pop(animated: true)
+            }
         }
         router.push(vc, animated: true, hideBottomBarWhenPushed: true)
     }
