+++ TOASTER-iOS/Present/AddLink/Model/SaveLinkModel.swift
