+++ TOASTER-iOS/Present/AddLink/ViewModel/AddLinkViewModel.swift
@@ -1,30 +1,36 @@
 //
-//  SelectClipViewModel.swift
+//  AddLinkViewModel.swift
 //  TOASTER-iOS
 //
-//  Created by Gahyun Kim on 2024/02/27.
+//  Created by Gahyun Kim on 9/19/24.
 //
 
 import Combine
 import UIKit
 
-final class SelectClipViewModel: ViewModelType {
+final class AddLinkViewModel: ViewModelType {
     
-    private var cancelBag = CancelBag()
+    private var cancelBag: CancelBag = CancelBag()
+    
     var selectedClip: [RemindClipModel] = []
-    
-    // MARK: - Input State
+    let embedLinkText = PassthroughSubject<String, Never>()
     
     struct Input {
+        let embedLinkText: AnyPublisher<String, Never>
+        let clearButtonTapped: AnyPublisher<Void, Never>
+        
         let requestClipList: Driver<Void>
         let clipNameChanged: Driver<String>
         let addClipButtonTapped: Driver<String>
         let completeButtonTapped: Driver<(String, Int?)>
     }
     
-    // MARK: - Output State
-    
     struct Output {
+        let isClearButtonHidden = PassthroughSubject<Bool, Never>()
+        let isNextButtonEnabled = CurrentValueSubject<Bool, Never>(false)
+        let textFieldBorderColor = PassthroughSubject<UIColor, Never>()
+        let linkEffectivenessMessage = PassthroughSubject<String?, Never>()
+        
         let needToReload = PassthroughSubject<Void, Never>()
         let duplicateClipName = PassthroughSubject<Bool, Never>()
         let addClipResult = PassthroughSubject<Bool, Never>()
@@ -32,10 +38,39 @@ final class SelectClipViewModel: ViewModelType {
         let navigateToLogin = PassthroughSubject<Void, Never>()
     }
     
-    // MARK: - Method
-    
     func transform(_ input: Input, cancelBag: CancelBag) -> Output {
         let output = Output()
+        
+        let inputText = input.embedLinkText
+            .merge(with: input.clearButtonTapped.map { "" })
+            .eraseToAnyPublisher()
+        
+        inputText
+            .map { $0.isEmpty }
+            .sink { isHidden in
+                output.isClearButtonHidden.send(isHidden)
+            }
+            .store(in: cancelBag)
+        
+        let isValid = inputText
+            .map { self.isValidURL($0) }
+            .share()
+            .eraseToAnyPublisher()
+        
+        isValid
+            .combineLatest(inputText.map { !$0.isEmpty })
+            .map { $0 && $1 }
+            .sink { isEnabled in
+                output.isNextButtonEnabled.send(isEnabled)
+            }
+            .store(in: cancelBag)
+        
+        input.embedLinkText
+            .map { $0.isEmpty ? "링크를 입력해주세요" : (self.isValidURL($0) ? nil : "유효하지 않은 형식의 링크입니다. " ) }
+            .sink { message in
+                output.linkEffectivenessMessage.send(message)
+            }
+            .store(in: cancelBag)
         
         input.requestClipList
             .networkFlatMap(self, { context, _ in
@@ -87,9 +122,19 @@ final class SelectClipViewModel: ViewModelType {
     }
 }
 
+private extension AddLinkViewModel {
+    func isValidURL(_ urlString: String) -> Bool {
+        if (urlString.prefix(8) == "https://") || (urlString.prefix(7) == "http://") {
+            return true
+        } else {
+            return false
+        }
+    }
+}
+
 // MARK: - Network
 
-private extension SelectClipViewModel {
+private extension AddLinkViewModel {
     func postSaveLink(url: String, category: Int?) -> AnyPublisher<Bool, Error> {
         return Future<Bool, Error> { promise in
             let request = PostSaveLinkRequestDTO(linkUrl: url, categoryId: category)
